Record menu: disallow names with spaces or hyphens.
Since these characters are used to determine concatenation etc in the
practice menu, if we have a name with a space or a hyphen the name will
be unable to be practiced - so disallow such names from being added to
the database in the first place. Catherine has stated previously that
all names supplied in her database will not have spaces/hyphens.