Add text wrapping to listen menu items.
If a very long name is present in the listen menu, the like and play
buttons are offset and are no longer visible for all names in the menu.
Adding text wrapping prevents this bug.

One side-effect is that the date text is now shifted over towards the
buttons. This effect however is minor and does not significantly impact
user experience.