Prevent audio clips from overlapping when playing.
[Untested]

A callback is now pased into {File,Memory}BackedAudioClip for when a new
ffplay process is created. The audio system handles this callback by
killing the previous ffplay process early if it existed. This means that
when audio is played, it will no longer overlap - the previously playing
audio will now be cut off.